refactor(signature): refactor client signature
A new function to generate client has been added to the ClientRepositoryInterface and ClientRepository. Implementation details of the signing process were moved from the Guard class to the Signer class. A new config for signature was also added. Changes include refactoring of some related classes and addition of new factory classes for creating the repository.